Location Attachment folder re-structure
