+++ DemoApp/StreamChat/Components/CustomAttachments/DemoChatMessageListVC.swift
@@ -1,0 +1,7 @@
+//
+// Copyright © 2023 Stream.io Inc. All rights reserved.
+//
+
+import StreamChatUI
+
+class DemoChatMessageListVC: ChatMessageListVC {}
