+++ DemoApp/StreamChat/Components/CustomAttachments/LocationAttachment/LocationAttachmentViewDelegate.swift
@@ -1,0 +1,19 @@
+//
+// Copyright © 2023 Stream.io Inc. All rights reserved.
+//
+
+import StreamChat
+import StreamChatUI
+
+protocol LocationAttachmentViewDelegate: ChatMessageContentViewDelegate {
+    func didTapOnLocationAttachment(
+        _ attachment: ChatMessageLocationAttachment
+    )
+}
+
+extension DemoChatMessageListVC: LocationAttachmentViewDelegate {
+    func didTapOnLocationAttachment(_ attachment: ChatMessageLocationAttachment) {
+        let mapViewController = LocationDetailViewController(locationAttachment: attachment)
+        navigationController?.pushViewController(mapViewController, animated: true)
+    }
+}
